+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/ObjectStore.java
@@ -126,11 +126,9 @@ import org.apache.hadoop.hive.metastore.model.MDBPrivilege;
 import org.apache.hadoop.hive.metastore.model.MDataConnector;
 import org.apache.hadoop.hive.metastore.model.MDCPrivilege;
 import org.apache.hadoop.hive.metastore.model.MDatabase;
-import org.apache.hadoop.hive.metastore.model.MDelegationToken;
 import org.apache.hadoop.hive.metastore.model.MFieldSchema;
 import org.apache.hadoop.hive.metastore.model.MFunction;
 import org.apache.hadoop.hive.metastore.model.MISchema;
-import org.apache.hadoop.hive.metastore.model.MMasterKey;
 import org.apache.hadoop.hive.metastore.model.MMetastoreDBProperties;
 import org.apache.hadoop.hive.metastore.model.MPackage;
 import org.apache.hadoop.hive.metastore.model.MPartitionEvent;
@@ -1988,195 +1986,6 @@ public class ObjectStore implements RawStore, Configurable {
       LOG.debug("Done executing cleanupEvents");
     }
     return delCnt;
-  }
-
-  private MDelegationToken getTokenFrom(String tokenId) {
-    try (QueryWrapper query = new QueryWrapper(pm.newQuery(MDelegationToken.class, "tokenIdentifier == tokenId"))) {
-      query.declareParameters("java.lang.String tokenId");
-      query.setUnique(true);
-      MDelegationToken delegationToken = (MDelegationToken) query.execute(tokenId);
-      return delegationToken;
-    }
-  }
-
-  @Override
-  public boolean addToken(String tokenId, String delegationToken) {
-
-    LOG.debug("Begin executing addToken");
-    boolean committed = false;
-    MDelegationToken token;
-    try{
-      openTransaction();
-      token = getTokenFrom(tokenId);
-      if (token == null) {
-        // add Token, only if it already doesn't exist
-        pm.makePersistent(new MDelegationToken(tokenId, delegationToken));
-      }
-      committed = commitTransaction();
-    } finally {
-      rollbackAndCleanup(committed, null);
-    }
-    LOG.debug("Done executing addToken with status : {}", committed);
-    return committed && (token == null);
-  }
-
-  @Override
-  public boolean removeToken(String tokenId) {
-
-    LOG.debug("Begin executing removeToken");
-    boolean committed = false;
-    MDelegationToken token;
-    try{
-      openTransaction();
-      token = getTokenFrom(tokenId);
-      if (null != token) {
-        pm.deletePersistent(token);
-      }
-      committed = commitTransaction();
-    } finally {
-      rollbackAndCleanup(committed, null);
-    }
-    LOG.debug("Done executing removeToken with status : {}", committed);
-    return committed && (token != null);
-  }
-
-  @Override
-  public String getToken(String tokenId) {
-
-    LOG.debug("Begin executing getToken");
-    boolean committed = false;
-    MDelegationToken token;
-    try{
-      openTransaction();
-      token = getTokenFrom(tokenId);
-      if (null != token) {
-        pm.retrieve(token);
-      }
-      committed = commitTransaction();
-    } finally {
-      rollbackAndCleanup(committed, null);
-    }
-    LOG.debug("Done executing getToken with status : {}", committed);
-    return (null == token) ? null : token.getTokenStr();
-  }
-
-  @Override
-  public List<String> getAllTokenIdentifiers() {
-    LOG.debug("Begin executing getAllTokenIdentifiers");
-    boolean committed = false;
-    Query query = null;
-    List<String> tokenIdents = new ArrayList<>();
-
-    try {
-      openTransaction();
-      query = pm.newQuery(MDelegationToken.class);
-      List<MDelegationToken> tokens = (List<MDelegationToken>) query.execute();
-      pm.retrieveAll(tokens);
-      committed = commitTransaction();
-
-      for (MDelegationToken token : tokens) {
-        tokenIdents.add(token.getTokenIdentifier());
-      }
-      return tokenIdents;
-    } finally {
-      LOG.debug("Done executing getAllTokenIdentifers with status : {}", committed);
-      rollbackAndCleanup(committed, query);
-    }
-  }
-
-  @Override
-  public int addMasterKey(String key) throws MetaException{
-    LOG.debug("Begin executing addMasterKey");
-    boolean committed = false;
-    MMasterKey masterKey = new MMasterKey(key);
-    try{
-      openTransaction();
-      pm.makePersistent(masterKey);
-      committed = commitTransaction();
-    } finally {
-      rollbackAndCleanup(committed, null);
-    }
-    LOG.debug("Done executing addMasterKey with status : {}", committed);
-    if (committed) {
-      return ((IntIdentity)pm.getObjectId(masterKey)).getKey();
-    } else {
-      throw new MetaException("Failed to add master key.");
-    }
-  }
-
-  @Override
-  public void updateMasterKey(Integer id, String key) throws NoSuchObjectException, MetaException {
-    LOG.debug("Begin executing updateMasterKey");
-    boolean committed = false;
-    Query query = null;
-    MMasterKey masterKey;
-    try {
-      openTransaction();
-      query = pm.newQuery(MMasterKey.class, "keyId == id");
-      query.declareParameters("java.lang.Integer id");
-      query.setUnique(true);
-      masterKey = (MMasterKey) query.execute(id);
-      if (null != masterKey) {
-        masterKey.setMasterKey(key);
-      }
-      committed = commitTransaction();
-    } finally {
-      rollbackAndCleanup(committed, query);
-    }
-    LOG.debug("Done executing updateMasterKey with status : {}", committed);
-    if (null == masterKey) {
-      throw new NoSuchObjectException("No key found with keyId: " + id);
-    }
-    if (!committed) {
-      throw new MetaException("Though key is found, failed to update it. " + id);
-    }
-  }
-
-  @Override
-  public boolean removeMasterKey(Integer id) {
-    LOG.debug("Begin executing removeMasterKey");
-    boolean success = false;
-    Query query = null;
-    MMasterKey masterKey;
-    try {
-      openTransaction();
-      query = pm.newQuery(MMasterKey.class, "keyId == id");
-      query.declareParameters("java.lang.Integer id");
-      query.setUnique(true);
-      masterKey = (MMasterKey) query.execute(id);
-      if (null != masterKey) {
-        pm.deletePersistent(masterKey);
-      }
-      success = commitTransaction();
-    } finally {
-      rollbackAndCleanup(success, query);
-    }
-    LOG.debug("Done executing removeMasterKey with status : {}", success);
-    return (null != masterKey) && success;
-  }
-
-  @Override
-  public String[] getMasterKeys() {
-    LOG.debug("Begin executing getMasterKeys");
-    boolean committed = false;
-    Query query = null;
-    List<MMasterKey> keys;
-    try {
-      openTransaction();
-      query = pm.newQuery(MMasterKey.class);
-      keys = (List<MMasterKey>) query.execute();
-      pm.retrieveAll(keys);
-      committed = commitTransaction();
-
-      String[] masterKeys = new String[keys.size()];
-      for (int i = 0; i < keys.size(); i++) {
-        masterKeys[i] = keys.get(i).getMasterKey();
-      }
-      return masterKeys;
-    } finally {
-      LOG.debug("Done executing getMasterKeys with status : {}", committed);
-      rollbackAndCleanup(committed, query);
-    }
   }
 
   // compare hive version and metastore version
