+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/metastore/iface/ConstraintStore.java
@@ -193,5 +193,4 @@ public interface ConstraintStore {
    * @throws MetaException error accessing the RDBMS
    */
   List<SQLCheckConstraint> addCheckConstraints(List<SQLCheckConstraint> cc) throws InvalidObjectException, MetaException;
-
 }
